dstack-py/requirements.txt
wheel
cloudpickle
deprecation
matplotlib
packaging
pandas
PyYAML>=5.3.1
requests
scikit-learn
tqdm
twine
torch
numpy